gpu: ion: Make ion_buffer_destroy return the number of bytes freed
Sometimes it can be useful to know how many bytes were actually freed
to the system allocator because of a call to ion_buffer_destroy. For
example, when we free some bytes from a shrinker function, it's
necessary for correct accounting to know whether the buffers we're
freeing are just going back onto a heap-specific pool (or similar) or
if they're actually going back to the system.

This also requires a change to the .free heap callback. Since only the
heaps know how many bytes were freed to the system allocator they must
return that value from their .free callback.

For now, don't return the bytes freed up by meta-information and other
housekeeping structures. Just report the freed bytes due to actual
buffer storage.

Change-Id: I62cb572d1c494432b303af05541accb76cd013ec
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>