video: adf: handle post-during-blank corner case
If a configuration is posted while one of the displays is off, ADF acts
as if the configuration was immediately posted to the display and then
removed.  In practice this policy eliminates annoying corner cases in
the client's display pipeline.

To support this, clients must list the target interfaces for each post.

Change-Id: Id9e0ebd88da5cc7fcfc506f612edb803b2cf5776
Signed-off-by: Greg Hackmann <ghackmann@google.com>