BACKPORT: powerpc/pseries: Remove (de)compression in nvram with pstore enabled
(cherry-pick from android-3.18 commit a3a5e94131d96ecdee7b1873a9d6215dc0ea7fff)

(De)compression support is provided in pstore in subsequent patches which
needs an additional argument 'compressed' to determine if the data
is compressed or not. This patch will take care of removing (de)compression
in nvram with pstore which was making use of 'hsize' argument in pstore write
as 'hsize' will be removed in the subsequent patch.

Signed-off-by: Aruna Balakrishnaiah <aruna@linux.vnet.ibm.com>
Reviewed-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Tony Luck <tony.luck@intel.com>
Bug: 31057326
Change-Id: I3d6ece25d2172e731b9a68ba7de5ab38da7930df