ion: fix dma APIs for compiling on non-ARM
__dma_page_cpu_to_dev is a private ARM api.  We can get the same
behavior by calling dma_sync_sg_for_device with a scatterlist
containing a single page.  It's still not quite a kosher use of
the dma apis, we still conflate physical addresses with bus
addresses, but it should at least compile on all platforms, and
work on any platform that doesn't have a physical to bus address
translation.

Change-Id: I8451c2dae4bf85841015c016640684ac28430a5a
Signed-off-by: Colin Cross <ccross@android.com>