gpu: ion: add page pooling to cached allocations from the system heap
Add a page pool for cached allocations from the system heap. This
results in a substantial performance boost.

Here are some quick results comparing latency of the ION_IOC_ALLOC ioctl
with and without page pooling for cached buffers (all times are in
milliseconds).

With a "debug" defconfig (CONFIG_SLUB_DEBUG, etc):

    | with pooling | without pooling |
    |--------------+-----------------|
    |        16.65 |           10.85 |
    |         0.04 |            8.26 |
    |         0.03 |            9.24 |
    |         0.02 |            9.78 |
    |         0.03 |           10.44 |
    |         0.09 |            9.25 |
    |         0.06 |            8.43 |
    |         0.04 |            9.83 |
    |         0.04 |           10.03 |
    |         0.05 |            8.32 |

With a "performance-tuned" defconfig:

    | with pooling | without pooling |
    |--------------+-----------------|
    |         1.01 |            3.78 |
    |         0.39 |            2.64 |
    |         0.04 |            4.13 |
    |         0.05 |            3.90 |
    |         0.04 |            2.59 |
    |         0.05 |            2.62 |
    |         0.05 |            3.92 |
    |         0.04 |            3.80 |
    |         0.04 |            3.54 |
    |         0.33 |            2.48 |

As illustrated by these results, the impact of not using pooling is
magnified by certain kernel configuration options. Page pooling is more
resilient to questionable configs. But even in the case of a
performance-tuned defconfig page pooling is faster by a few orders of
magnitude, on average.

Change-Id: I6d7688026ff4ee65efa047a164652da1427e72c4
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>