audit: Add cmdline to taskinfo output
On some devices, the cmdline and task info vary. For instance, on
Android, the cmdline is set to the package name, and the task info
is the name of the VM, which is not very helpful.

The additional cmdline output only runs if the audit feature
is set dynamically at run time.

Change-Id: I98a417c9ab3b95664c49aa1c7513cfd8296b6a2a

Signed-off-by: William Roberts <wroberts@tresys.com>