[RFC] ion: Add carveout and chunk heaps to dummy driver
Add support to the dummy driver for basic carveout and chunk heaps.

Since we're generating these heaps at module_init, and we want
this driver to be generic enough to be tested on any arch, we
don't have the ability to alloc bootmem, so both of these heaps
are conventionally allocated using kzalloc(), which limits us
to 4M in size.

XXX: Not for merging, just for review!  This currently has some
issues on ARM, but I've managed to make both heaps work on x86_64
w/ Colins' patches and some additional fix ups of my own to the
ION core and heap code.

Change-Id: I5b973ffa38c134e08360df701823db3206b31255
Signed-off-by: John Stultz <john.stultz@linaro.org>