[RFC] ion: Add carveout and chunk heaps to dummy driver
Add support to the dummy driver for basic carveout and chunk heaps.

Since we're generating these heaps at module_init, and we want
this driver to be generic enough to be tested on any arch, we
don't have the ability to alloc bootmem, so both of these heaps
are conventionally allocated using alloc_pages(), which limits us
to 4M in size.

This version resolves earlier problems with ARM by using
alloc_pages_exact() instead of kzalloc().

XXX: Still need to look into using CMA for heap allocation here.

Change-Id: I5b973ffa38c134e08360df701823db3206b31255
Signed-off-by: John Stultz <john.stultz@linaro.org>