staging: android: ashmem: add new ASHMEM_GET_PURGED ioctl command
This new ioctl command lets userspace retrieve the ranges that got purged in an
ashmem region.

It can be used by userspace in order to minimize fragmentation in the address
space by reusing the ranges that got purged (without having to lock them to
know whether they got purged) instead of creating new ashmem regions.

In case a lot of regions got purged the ioctl may have to be called in a loop
from userspace since the kernel is using a buffer of a limited size. Note that
the ioctl returns the number of ranges that were written by the kernel to the
buffer provided by userspace.

Signed-off-by: Philippe Liard <pliard@google.com>