Update patch set 1
Patch Set 1:

I'm not a big fan of adding new apis to ashmem.  Direct access to ashmem is not a platform feature, and trying to access it directly is a recipe for trouble - it is likely to be deprecated and replaced by "volatile ranges" by the upstream kernel at some point in the future.  In addition, any changes to ashmem will not be widely available unless you add a CTS test to force vendors to pull in your patch, which adds a bunch of work for them.

The expected use case for ashmem is when you are storing data you know you'll need again, but can recreate if necessary.  In that use case, there is no reason to check if a page has been purged without pinning it.

This looks like you are trying to use it to hold a cache, and let the kernel kill cache entries for you.  I don't think that is going to result in good behavior - you have no control over how fast the kernel ages your cache relative to anything else, or in what order it ages your cache.  ashmem is already a little unpredictable - it is likely to do nothing for a long time, and then suddenly wipe out all of your memory.

I would expect a cache to want to listen for a memory pressure signal, and then use that to free pages out of the cache in an order that makes sense to the app.  We already have some memory pressure signals that use app deaths from the lowmemorykiller as a signal that apps higher up the LRU list should free some memory.  If those signals are not good enough we should work to improve them, and new kernels have a mempressure signal that could help.  I don't think we should be exposing ashmem details like this all the way up to apps.

Patch-set: 1