Update patch set 1
Patch Set 1:

Thanks Colin and John,

Colin, it's definitely possible that using the current/new implementation of the signals would be good enough. In the past this hasn't been enough to prevent us from being killed by the OOM killer though. How strong would the guarantees be now? Of course it also depends I guess on our ability in userspace to react to those signals quickly.

In general we are more interested in not being killed than having a fine-tuned eviction logic. Just like you mentioned we also noticed that ashmem eviction was rather nonexistent or quite aggressive :) 

Let me explain our experience with ashmem so that you have all the context.

We have been using ashmem for quite a while now in Chrome to store the decoded bitmaps (that can be as large as 10 MBytes). We are also looking into using it for more types of resource (like the resources sent by the server that we keep not only in the HTTP disk cache but also memory resident).

We used to have a very limited use of ashmem due to the process file descriptor limit so we were allowing only up to 128 ashmem regions (which was quite conservative) to avoid running out of file descriptors. When we were exceeding this soft limit (which turned out to happen quite often) we were falling back to heap allocations and the heap-allocated bitmaps were stored in a separate LRU cache.

We are now looking into scaling up the use of discardable memory in Chrome. The first problem was the fact that ashmem was a limited resource due to the file descriptor limit if it was used in a naive way (like we were doing in Chrome).
So we are now using an allocator that operates on top of large (e.g. 64 MBytes) ashmem regions so that we use only a couple of file descriptors (see http://src.chromium.org/viewvc/chrome/trunk/src/base/memory/discardable_memory_allocator_android.cc). Thanks to this we can now do as many ashmem allocations as we need (of course ashmem should still only be used for relatively large allocations). This allows our renderer processes (in which decoded images are stored) not to be killed by the OOM killer as they were before due to an excessive use of non-evictable physical memory.

The problem we have now is that we are sometimes running out of address space. Let me explain. When we were creating individual ashmem regions for each bitmap (i.e. before we start using the allocator) we were also calling munmap() every time we were calling ashmem_unpin_region() (and mmap() every time we were calling ashmem_pin_region()). This was effectively protecting us from running out address space when we were keeping a few big (unpinned) ashmem regions around.
Unfortunately we can no longer do that with the allocator since it relies on the fact that the ranges inside the large ashmem regions it operates on are contiguous.
Of course we can also fallback to allocating individual ashmem regions (that we would be able to munmap()) when there is address space pressure (assuming there is a good way to detect "address space pressure"). But it doesn't really scale either since we would be back to using file descriptors.

Just to be clear (and I should have been clearer earlier :)), we would not have had this problem in the first place if our eviction logic in our bitmap cache wasn't buggy. We were allowing up to 128 bitmaps stored in discardable memory before our LRU cache (designed in the first place for bitmaps allocated in the heap) started accounting memory usage (and a bitmap can be quite large as I mentioned above).

On the other hand ashmem does quite a good job at evicting the unpinned ranges (although when eviction happens, literally all the unpinned ranges get evicted as Colin mentioned). This means that we are running out of address space when all our unpinned ashmem regions got purged by the kernel which is quite unfortunate since we could reuse them (meaning that we would use less large ashmem regions in general therefore we would cause much less pressure in the address space). This means that we would be doing in userspace some sort of garbage collection of the purged ranges when we would be exceeding a certain amount of virtual memory. This is what this patch was about.


John, my (limited :)) understanding is that whether this problem could happen with volatile ranges boils down to this question: Do you expect userspace to have to roll its own (non-broken) LRU cache on top of volatile ranges?

- If you do then I don't see any problem with volatile ranges. Userspace would have a hard limit in its LRU cache preventing it from running out of address space. Having a limit in userspace can be non-{ideal,trivial} IMO though.

- If you expect userspace to be able to completely rely on the kernel for the eviction of the unpinned ranges then it depends on whether userspace could munmap() its unpinned ranges, right?

* If volatile ranges use file descriptors then this means that userspace would have to roll its allocator on top of volatile ranges (like we did with ashmem) which would make unmapping non-trivial/impossible?

* If volatile ranges don't use file descriptors then userland might not have to use a custom allocator which would make munmap()/mmap() possible upon unpinning/pinning?
On the performance side there is also the fact that mmap() and munmap() are not free. If userspace has to invoke them for every single unpin/pin then it would be slightly unfortunate IMO (although this could be done by userspace only when there is pressure in the address space. Also note the fact that although userspace would not have to roll its allocator if volatile ranges don't use file descriptors, it might still want to do it for performance reasons to minimize syscalls depending on how expensive creating a volatile range is.

* So in case munmap() is not possible (or not desirable for performance reasons) then I think a feature such as the one suggested by this patch might be valuable.


I hope this makes any sense :) What do you think?

Thanks!

Patch-set: 1
Reviewer: Philippe Liard <1019003@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0