ion: support multiple address space map
Ion doesn't support multiple address space. The recent IOMMU usually
provide multiple address spaces, where a mapping depends on an address
space. An address space is bind to a device pointer passed from
attachment. Those info needs to be stored in a buffer info. If the
same buffer needs to be mapped into the same AS, the same mapping info
needs to be passed back.

Change-Id: I8445fe36a6d40c89d1fcd4e63b761a7cb458f87f
Signed-off-by: Lauri Peltonen <lpeltonen@nvidia.com>
Reviewed-by: Hiroshi Doyu <hdoyu@nvidia.com>
Signed-off-by: Allen Martin <amartin@nvidia.com>
Signed-off-by: Andrew Chew <achew@nvidia.com>