nf: IDLETIMER: time-stamp and suspend/resume handling.
Message notifications contains an additional timestamp field in nano
seconds.
The expiry time for the timers are modified during suspend/resume.
If radio goes down while the system is suspended then
a notification is sent when it resumes with the past timestamp.

Removes the race condition for multiple work scheduled.

Change-Id: I367b40209b4f8ceff5b60c13fce27bb4b78b72da
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>
Bug: 13247811