ARM: Fix ldrt fixup handler for und_usr faults
The exception handler for undefined intruction accesses in
user mode executes ldrt or ldrht instructions depending on
whether the instruction is executed in ARM or THUMB mode.
Since these two instructions can cause data aborts, the
current implementation has setup a fixup handler for the
instruction which simply ignores the instruction and jumps
directly to ret_from_exception. This handler works fine for
invalid data accesses since the application is likely to hit
the same abort on the next instruction and exit gracefully.
However, if the exception is caused by a valid page fault,
the page is fetched on the next instruction but the instruction
causing the fault is silently ignored and the application
continues executing the following instruction.

The fixup exception handler should be modified to re-execute
the faulting instruction so any necessary steps required to
handle the exception are taken correctly. If the exception
is caused by a page fault, the page is re-mapped, and the
faulting instruction is successfully re-executed.

The logic modifies the fixup handler to adjust the value of the
PC back to the faulting instruction, depending on whether the
instruction was executed in ARM or THUMB mode, before returning
to ret_from_exception.

b/13695295

Signed-off-by: Danesh Petigara <dpetigara@broadcom.com>
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>