Update patch set 2
Patch Set 2: Code-Review+1

JP: This patch addresses long standing confusion about SELinux denials. In particular, Android developers can't tell if an SELinux denial is just a warning (because the domain is in permissive mode) or if it's actually enforced (because the domain is in enforcing). This patch allows us to distinguish between the two cases.

Patch-set: 2
Reviewer: Nick Kralevich <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1