Update patch set 1
Patch Set 1:

There are two SAR sensors on the LTE version(?) how are we handling that? I expect there may have to be SAR driver changes, please create a bug against the SAR driver to function as per the defacto standard you are expecting and I will ensure that Tim complies.

Patch-set: 1