Update patch set 3
Patch Set 3:

Noticed the following while verifying the 3.10 gerrit:
Observed that on a ADK2.0 kit, the Userspace was not configuring sys.usb.config to accessory,audio_source,adb. Not sure whether kernel was sending the ACCESSORY START uevent back to userspace. Target works fine without the change. Was this verified on the target ?

Patch-set: 3
Label: Verified=0