net: ipv6: Add a sysctl to treat optimistic addresses as preferred
Add a sysctl that causes an interface's optimistic addresses
to be considered equivalent to other non-deprecated addresses
for source address selection purposes.  Preferred addresses
can still take precendence over optimistic addresses though.

This is useful where different interfaces are connected to
different networks from different ISPs (e.g., a cell network
and a home wifi network).

The current behaviour complies with RFC 3484/6724, and it
makes sense if the host has only one interface, or has
multiple interfaces on the same network, but not in the
multiple distinct networks case.

Also: add an entry in ip-sysctl.txt for optimistic_dad.

Change-Id: Ic7e50781c607e1f3a492d9ce7395946efb95c533
Signed-off-by: Erik Kline <ek@google.com>