Update patch set 6
Patch Set 6:

(5 comments)

Code LGTM. I think the commit message and documentation could be a bit clearer. Particularly, I've suggested an example to put in the commit message to clearly show the problem. I've seen commit messages for networking changes that are extremely verbose (even to the point of containing test results/benchmarks etc.), so I don't think there's a risk of it getting too long.

Once we're happy with this we should send it to netdev (based off of net-next), to see if upstream will merge it.

Patch-set: 6
Label: Verified=0