Update patch set 1
Patch Set 1:

IIUC, this is not necessary and won't actually help you with being able to set security.selinux attributes on these inodes.  The vfs already provides a fallback behavior for the security.* attributes such that if the filesystem does not implement the xattr handlers, it will call directly into the security module (i.e. SELinux).  This suffices as long as the attribute does not require persistent storage (which your patch does not provide anyway) and the inode is pinned in memory for its lifetime (unlike e.g. sysfs, where there is a separate backing data structure where we must save the attribute in order to restore it when the inode is evicted and later re-created).  What you may want is a patch along the lines of what we did for rootfs and sysfs in security/selinux/hooks.c, where we leave them as genfscon in the policy to provide a policy default value but permit them to be set via setxattr from userspace to more specific values.  Just look for references to rootfs or sysfs in that file.  We need to generalize this as obviously it is ugly to keep embedding hardcoded filesystem type names in SELinux, but that will be more involved (extension to the policy to permit specifying a new form of inode labeling).

Patch-set: 1
Label: Code-Review=0