selinux: add ioctl command SID cache
Add capability to whitelist ioctl command numbers in policy.

Much like the network port/address/interface cache on which this design is
based, this adds an ioctl command cache. The design is almost completely
based off the netport cache. The mapping of ioctl command to SID is stored
in a hash table. Ioctl commands are composed of four values stored in an
unsigned integer. Using the least significant bits of each of these four
values achieves even spreading across buckets.

Bug: 18087110
Signed-off-by: Jeff Vander Stoep <jeffv@google.com>