selinux: optional whitelisting of ioctl command numbers
Patch maintains the existing file descriptor check on ioctls and adds
an optional check on the command number. Policydb version number has
been incremented to POLICYDB_VERSION_IOCTL_OPERATIONS=30, and the
av_decision structure modified to support an array of allowed ioctl
command ranges that are specified in policy. Only the type and number
bitfields of the ioctl command are considered, the size and direction
bitfields are ignored. This patch does not address the difference in
command numbers used on a subset of the supported architectures e.g.
mips.

Bug: 18087110
Signed-off-by: Jeff Vander Stoep <jeffv@google.com>
Change-Id: I0d9d014a7fd66db4912fc583a5414d93854ecc14