subsystem:Set cpu_load calculation on current freq
We are doing calcuation in timer callback on target_freq.
If cpu frequency driver fails to set the freq to target_freq, the cpu_load
calculation is incorrect which is calculating on target_freq.
The cpu_load and other calculation should be done on the
pcpu->policy->cur because CPU is working on target frequency.
The cpu delta_time and delta_idle is also working current freq.

Signed-off-by: rahul.khandelwal <rahul.khandelwal@spreadtrum.com>