net: core: fix Null ptr dereference in UID-based routing
We call sock_i_uid(struct sock *) without checking for a valid
"struct sock" pointer, resulting in a Null pointer dereference.

Not that we need to check for valid pointer everytime we call
sock_i_uid() but plan is to at least do it when we are doing
this Null pointer check for other arguments.

Signed-off-by: Amit Pundir <amit.pundir@linaro.org>