Update patch set 1
Patch Set 1:

> Stephen are you asking for this change to get into the kernel.org
 > maintenance tree for 3.14?

No.  Just noting that when/if this change is downmerged to any device kernel tree, it depends on the other change (in order to build; it isn't truly a logical dependency, just that the prior change made a change to an interface used by this change).  Neither change is really suitable for stable, as they do not fix any bugs.

Patch-set: 1