Update patch set 1
Patch Set 1:

The set of options included (and not included) here were discussed in review of the original change for android-3.4,
https://android-review.googlesource.com/#/c/89730/

Would like to specify values for all of the SELinux related options but merge_config.sh complains if you specify a value and it is left unspecified in the defconfig file, even if the default is the value you wanted.  That was also discussed in the comments of the above change.

Patch-set: 1