Update patch set 1
Patch Set 1: Code-Review+2

These turn into suggestions for partners to use as a starting-point. We need to request an audit by all the Nexus kernel maintainers to make sure these are set accordingly in the defconfig, if not already. @nnk@ do this like you have done all of the merge requests (open a bug, and let it rotate through the maintainers).

Patch-set: 1
Label: Code-Review=+2