Update patch set 1
Patch Set 1:

I don't think the protocol should be changed this way. If you want more detailed information we should add an error code as a parameter (like BR_ERROR), not add a new return type that the user-space library have to handle for every error we want to report.

Patch-set: 1