Fix BR_ERROR usage and change LSM denials to use it.
BR_ERROR specifies that it provides an error parameter, but it is not actually
used when errors are copied into the read buffer of for threads in
binder_thread_read().  Add a place to store BR_ERROR parameters as part of the
binder_thread struct and report these values when copying BR_ERROR into the read
buffer.  Also switch the LSM denials to make use of this to report -EPERM.

Bug: 18641854
Signed-off-by: Dan Cashman <dcashman@google.com>