firmware: Add android firmware path
Android firmware path is /vendor/firmware.  Add to the firmware search
paths.

If the kernel is able to find firmware in the kernel firmware search path
(when requested through request_firmware()) it will load it directly without
the involvement of the usermodehelper/ udev process (which is not always
available).

Derived from work by Neil Gabriel <ngabriel@nvidia.com> and Dan Willemsen
<dwillemsen@nvidia.com>.

Change-Id: I7e29de454e9dffcb4a50eee84677817d57eb3b1f
Signed-off-by: Andrew Chew <achew@nvidia.com>