RFC: Revert "ARM64: add option to build Image.gz/dtb combo"
This reverts commit 082d89f7f2ae529e2d7b50f6259e4430472cf4ed.

We ran into problems with this patch when trying to backport
patches that split arm64 dts into vendor directories. In
particular, make mrproper broke:
https://code.google.com/p/chromium/issues/detail?id=480825#c10

Now, this can probably be fixed, e.g. using this patch:
https://chromium-review.googlesource.com/#/c/272362/

However, I'm wondering if the best approach may not be to
drop this kernel option. This commit message:
https://chromium-review.googlesource.com/#/c/272909/
mentions that:
""
Upstream has taken a hard stance on not allowing custom image production
as part of the kernel build, so keeping the same policy in our downstream
project helps reduce confusion and makes our build environment agnostic
of where the kernel tree is coming from (and whether it has our local
changes to directly produce the Image.fit).
""

I'm not sure what Image.gz/dtb combos are used for, and if it is
easy to replace the kernel config option by some build-system
scripts.

Change-Id: I8dd15229402a92a28143713205dbd48f3c7b5784
Signed-off-by: Nicolas Boichat <drinkcat@google.com>